+++ src/classes/api/index.ts
@@ -148,6 +148,75 @@ export async function addNote(title: string, content: string, startDate: SimpleC
 }
 
 /**
+ * Add a custom button to the right side of the calendar (When in full view). The button will be added below the Note Search button.
+ *
+ * @param buttonTitle The text that appears when the button is hovered over.
+ * @param iconClass The Font Awesome Free icon class to use for the buttons display.
+ * @param customClass A custom CSS class to add to the button.
+ * @param showSidePanel If the button should open a side panel or not. A side panel functions like the notes list but will be completely empty.
+ * @param onRender Function that is called to show information to users, see below for more details.
+
+ * If the showSidePanel parameter is true this function will be passed an HTMLElement representing the side panel.
+ * The element could potentially be null so code should account for that.
+ * This function is then responsible for populating the side panel with any information.
+ * This function is called everytime the calendar is rendered.
+ *
+ * If the showSidePanel parameter is false this function will be passed an Event for the click.
+ * The event could potentially be null so code should account for that.
+ * This function should then open up an application or dialog or perform an action for the user.
+ * This function is called everytime the button is clicked.
+ *
+ * @example
+ * ```javascript
+ * // Function to call to populate the side panel
+ * function populateSidePanel(event, element){
+ *     if(element){
+ *            const header = document.createElement('h2');
+ *            header.innerText = "My Custom Button";
+ *            element.append(header);
+ *        }
+ * }
+ *
+ * // Function to call when the button is clicked
+ * function sidePanelButtonClick(event, element){
+ *     if(event){
+ *          const dialog = new Dialog({
+ *             title: "My Module",
+ *             content: "You clicked the button!",
+ *             buttons:{
+ *                 awesome: {
+ *                     icon: '<i class="fa-solid fa-face-smile"></i>',
+ *                     label: "Awesome!"
+ *                 }
+ *             },
+ *             default: "awesome"
+ *         });
+ *         dialog.render(true);
+ *     }
+ * }
+ *
+ *
+ * // Adding a button that should show a side panel
+ * // Clicking the button will show a side panel that will have the title "My Custom Button"
+ * SimpleCalendar.api.addSidebarButton("My Module", "fa-computer-mouse", "my-custom-class", true, populateSidePanel);
+ *
+ * // Adding a button that will show a dialog when clicked
+ * SimpleCalendar.api.addSidebarButton("My Module", "fa-computer-mouse", "my-custom-class", false, sidePanelButtonClick);
+ *
+ * ```
+ */
+export function addSidebarButton(buttonTitle: string, iconClass: string, customClass: string, showSidePanel: boolean, onRender: (event: Event | null, renderTarget: HTMLElement | null | undefined) => void){
+    MainApplication.addonButtons.push({
+        title: buttonTitle,
+        iconClass: iconClass,
+        customClass: customClass,
+        showSidePanel: showSidePanel,
+        onRender: onRender
+    });
+    MainApplication.updateApp();
+}
+
+/**
  * Advance the date and time to match the next preset time.
  *
  * **Important**: This function can only be run by users who have permission to change the date in Simple Calendar.