+++ src/constants.ts
@@ -309,7 +309,8 @@ export enum SimpleCalendarHooks {
     DateTimeChange = 'simple-calendar-date-time-change',
     ClockStartStop = 'simple-calendar-clock-start-stop',
     PrimaryGM = 'simple-calendar-primary-gm',
-    Ready = 'simple-calendar-ready'
+    Ready = 'simple-calendar-ready',
+    Init = 'simple-calendar-init'
 }
 
 /**